+++ lib/presentation/screens/dashboard/panoramic_heatmap_screen.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:intl/intl.dart';
+import '../../../domain/services/ledger_service.dart';
 import '../../providers/panoramic_heatmap_provider.dart';
 import '../../../domain/models/heatmap_data.dart';
 import 'package:centim/l10n/app_localizations.dart';